# backend/src/main/java/com/kiezmarkt/listing/domain/ListingStatus.java
package com.kiezmarkt.listing.domain;

/**
 * Lifecycle status of a {@link Listing}. {@code deleted} is terminal.
 */
public enum ListingStatus {
    draft,
    published,
    paused,
    expired,
    deleted
}