# contracts/api.yaml
openapi: 3.1.0
info:
  title: Kiezmarkt Listing Service
  version: 0.1.0
  summary: Listings and their lifecycle, search, categories, saved searches.
  description: |
    The listing service surface of Kiezmarkt, the synthetic neighbourhood
    classifieds marketplace used as the guiding project of the AI-enablement
    training.

    The domain model behind this contract is `docs/project/domain.md`. Where the
    two disagree, `domain.md` wins and this file is the defect. Two invariants
    are carried by the schemas rather than by prose:

    - Money is an integer number of cents. No property in this document is typed
      `number`, at any depth.
    - No response schema exposes a seller email. A seller is visible as
      `sellerId`, and where a human-readable name is required, through the
      `SellerPublic` projection.

    `domain.md` specifies no authentication or authorization model. Endpoints
    that are inherently caller-scoped (`/saved-searches`) say so in their
    description; the mechanism is out of scope for this contract.
servers:
  - url: https://api.kiezmarkt.example/v1
    description: Synthetic. Nothing is deployed behind this URL.
...
paths:
  /listings:

    get:
      operationId: searchListings
      tags: [listings]
      summary: Search listings
      description: |
        Returns published listings only. A listing in `draft`, `paused`,
        `expired` or `deleted` is never part of a search result, regardless of
        filters (invariant 1 in `domain.md`).

        Pagination is cursor-based. A caller passes `limit` and, from the second
        page on, the `nextCursor` returned by the previous page. Cursors are
        opaque; a client must not construct, parse or persist them beyond the
        life of one result set.

        `domain.md` leaves open whether ordering is stable across pages when
        listings change status mid-pagination. This contract does not resolve
        that question; it is deliberately unanswered.
      parameters:
        - name: q
          in: query
          description: Free-text query over `title` and `description`.
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: categoryId
          in: query
          description: |
            Restrict to one category. A top-level id widens the search to all of
            its leaves; a leaf id matches exactly. Top-level categories never
            carry listings themselves (invariant 4).
          required: false
          schema:
            $ref: '#/components/schemas/CategoryId'
        - name: priceMinCents
          in: query
          description: Inclusive lower bound, in integer cents.
          required: false
          schema:
            $ref: '#/components/schemas/MoneyCents'
        - name: priceMaxCents
          in: query
          description: Inclusive upper bound, in integer cents.
          required: false
          schema:
            $ref: '#/components/schemas/MoneyCents'
        - name: postcode
          in: query
          description: Centre of the geographic filter.
          required: false
          schema:
            $ref: '#/components/schemas/Postcode'
        - name: radiusKm
          in: query
          description: |
            Radius around `postcode`, in kilometres. Requires `postcode`; sent
            alone it is a `400` with problem type `radius-without-postcode`.
          required: false
          schema:
            type: integer
            minimum: 1
        - name: cursor
          in: query
          description: Opaque cursor from the previous page's `nextCursor`.
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Page size.
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: A page of published listings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingPage'
              examples:
                firstPage:
                  summary: First page, two of 120 seeded listings
                  value:
                    items:
                      - id: lst_9f2c41
                        sellerId: sel_0142
                        title: Technics SL-1210 MK2 Plattenspieler
                        description: Gewartet, neue Nadel, Originalhaube vorhanden.
                        categoryId: elektronik.audio
                        priceCents: 42000
                        currency: EUR
                        status: published
                        postcode: '10437'
                        createdAt: '2026-08-11T07:41:12Z'
                        publishedAt: '2026-08-11T08:02:55Z'
                        imageIds: [img_3a1f, img_3a20, img_3a21]
                      - id: lst_20b7e3
                        sellerId: sel_0007
                        title: Umzugskartons, 12 Stueck
                        description: Zu verschenken, nur Abholung Neukoelln.
                        categoryId: haushalt.moebel
                        # `priceCents: 0`. `domain.md` records that both readings of 0 exist in
                        # the data (free vs. unpriced) and does not decide between them.
                        priceCents: 0
                        currency: EUR
                        status: published
                        postcode: '12047'
                        createdAt: '2026-08-14T16:20:03Z'
                        publishedAt: '2026-08-14T16:20:44Z'
                        imageIds: [img_77c0]
                    nextCursor: eyJvIjoyMCwidCI6IjIwMjYtMDgtMTQifQ
                lastPage:
                  summary: Last page, no further cursor
                  value:
                    items: []
                    nextCursor: null
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      operationId: createListing
      tags: [listings]
      summary: Create a listing
      description: |
        A created listing always starts in `status: draft` with
        `publishedAt: null`. The request body carries neither field; a body that
        does is rejected with `400`. Publishing is a separate call to
        `POST /listings/{id}/status`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingCreate'
            examples:
              minimal:
                summary: Only the required fields
                value:
                  sellerId: sel_0142
                  title: Bierkiste Sitzbank, selbst gebaut
                  categoryId: haushalt.moebel
                  priceCents: 2500
                  postcode: '10119'
      responses:
        '201':
          description: Created. The listing is a draft.
          headers:
            Location:
              description: URI of the created listing.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
              examples:
                created:
                  value:
                    id: lst_c40a19
                    sellerId: sel_0142
                    title: Bierkiste Sitzbank, selbst gebaut
                    description: ''
                    categoryId: haushalt.moebel
                    priceCents: 2500
                    currency: EUR
                    status: draft
                    postcode: '10119'
                    createdAt: '2026-08-31T09:14:00Z'
                    publishedAt: null
                    imageIds: []
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /listings/{id}:

    parameters:
      - $ref: '#/components/parameters/ListingId'

    get:
      operationId: getListing
      tags: [listings]
      summary: Read one listing
      description: |
        Returns a listing in any status except `deleted`, which is reported as
        `410`. Unlike `GET /listings`, this endpoint is not restricted to
        published listings: a seller reads their own draft here.
      responses:
        '200':
          description: The listing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
              examples:
                published:
                  summary: Published listing with a price
                  value:
                    id: lst_9f2c41
                    sellerId: sel_0142
                    title: Technics SL-1210 MK2 Plattenspieler
                    description: Gewartet, neue Nadel, Originalhaube vorhanden.
                    categoryId: elektronik.audio
                    priceCents: 42000
                    currency: EUR
                    status: published
                    postcode: '10437'
                    createdAt: '2026-08-11T07:41:12Z'
                    publishedAt: '2026-08-11T08:02:55Z'
                    imageIds: [img_3a1f, img_3a20, img_3a21]
                freeItem:
                  summary: Published listing at zero cents
                  value:
                    id: lst_20b7e3
                    sellerId: sel_0007
                    title: Umzugskartons, 12 Stueck
                    description: Zu verschenken, nur Abholung Neukoelln.
                    categoryId: haushalt.moebel
                    priceCents: 0
                    currency: EUR
                    status: published
                    postcode: '12047'
                    createdAt: '2026-08-14T16:20:03Z'
                    publishedAt: '2026-08-14T16:20:44Z'
                    imageIds: [img_77c0]
                # The example below does NOT validate against `Listing`. It is
                # reality, not the contract: 7 of the 120 seeded listings arrive
                # from the import feed with `priceCents: null` and are served
                # unchanged. Kept here on purpose so the gap between the declared
                # schema and the payload is visible in the contract itself. Do not
                # relax the schema to make it validate.
                unpricedFromFeed:
                  summary: Feed-imported listing with no price (schema-violating, real)
                  value:
                    id: lst_5e8802
                    sellerId: sel_0031
                    title: Sofa 3-Sitzer, nur Abholung
                    description: Zu verschenken. Kleiner Riss an der Armlehne.
                    categoryId: haushalt.moebel
                    priceCents: null
                    currency: EUR
                    status: published
                    postcode: '13355'
                    createdAt: '2026-07-29T11:03:41Z'
                    publishedAt: '2026-07-29T11:03:41Z'
                    imageIds: [img_91aa, img_91ab]
                draft:
                  summary: Draft, never published
                  value:
                    id: lst_c40a19
                    sellerId: sel_0142
                    title: Bierkiste Sitzbank, selbst gebaut
                    description: ''
                    categoryId: haushalt.moebel
                    priceCents: 2500
                    currency: EUR
                    status: draft
                    postcode: '10119'
                    createdAt: '2026-08-31T09:14:00Z'
                    publishedAt: null
                    imageIds: []
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'

    patch:
      operationId: updateListing
      tags: [listings]
      summary: Partially update a listing
      description: |
        Absent properties are left unchanged. No patchable property is nullable,
        so an explicit `null` is a `400` rather than a clear.

        `status` is not patchable here. It moves only through
        `POST /listings/{id}/status`, so that every transition passes one guard
        and emits one event. `id`, `sellerId`, `createdAt` and `publishedAt` are
        server-owned and rejected in a patch body.

        A `deleted` listing is not patchable (`409`, `terminal-status`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingPatch'
            examples:
              priceCorrection:
                summary: Correct a price
                value:
                  priceCents: 38000
              reorderImages:
                summary: Promote a different cover image
                value:
                  imageIds: [img_3a20, img_3a1f, img_3a21]
      responses:
        '200':
          description: The updated listing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /listings/{id}/status:

    parameters:
      - $ref: '#/components/parameters/ListingId'

    post:
      operationId: changeListingStatus
      tags: [listings]
      summary: Transition a listing to another status
      description: |
        The body carries the target status. The service either applies the
        transition and returns the listing, or rejects it with `409` and problem
        type `illegal-status-transition`.

        Legal transitions:

        | From | To |
        | :--- | :--- |
        | `draft` | `published`, `deleted` |
        | `published` | `paused`, `expired`, `deleted` |
        | `paused` | `published`, `expired`, `deleted` |
        | `expired` | `published`, `deleted` |
        | `deleted` | nothing. `deleted` is terminal (invariant 5) |

        Everything not in that table is illegal, including `draft` to `paused`,
        `draft` to `expired`, and any transition out of `deleted`. A transition
        to the status a listing already holds is also a `409`
        (`illegal-status-transition`), not a silent success, so that a caller
        cannot mistake a no-op for a state change.

        `publishedAt` is set on the first transition into `published` and never
        written again (invariant 2). A later `paused` to `published` leaves the
        original `publishedAt` in place.

        Every accepted transition is recorded in the service's status history.
        Rejected ones record nothing. Status changes are not published as
        events; see `docs/project/contracts/events.md`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusChangeRequest'
            examples:
              publish:
                summary: Publish a draft
                value:
                  status: published
                  reason: seller-action
              expire:
                summary: Expire a published listing
                value:
                  status: expired
                  reason: ttl-reached
      responses:
        '200':
          description: Transition applied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
              examples:
                published:
                  summary: Draft published, `publishedAt` set
                  value:
                    id: lst_c40a19
                    sellerId: sel_0142
                    title: Bierkiste Sitzbank, selbst gebaut
                    description: Robust, zwei Kisten, Sitzflaeche lackiert.
                    categoryId: haushalt.moebel
                    priceCents: 2500
                    currency: EUR
                    status: published
                    postcode: '10119'
                    createdAt: '2026-08-31T09:14:00Z'
                    publishedAt: '2026-08-31T09:31:07Z'
                    imageIds: [img_b110]
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            The transition is not legal from the listing's current status.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                illegalTransition:
                  summary: "`draft` to `paused`"
                  value:
                    type: https://kiezmarkt.example/problems/illegal-status-transition
                    title: Illegal status transition
                    status: 409
                    detail: Listing lst_c40a19 is in status draft; draft may only move to published or deleted.
                terminal:
                  summary: Anything out of `deleted`
                  value:
                    type: https://kiezmarkt.example/problems/illegal-status-transition
                    title: Illegal status transition
                    status: 409
                    detail: Listing lst_5e8802 is deleted. deleted is terminal.
...
components:
  parameters:
    ListingId:
      name: id
      in: path
      required: true
      description: Opaque, server-assigned listing id.
      schema:
        $ref: '#/components/schemas/ListingId'
...
  responses:
    BadRequest:
      description: The request is malformed or violates a field constraint.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            titleTooShort:
              value:
                type: https://kiezmarkt.example/problems/validation-failed
                title: Validation failed
                status: 400
                detail: title must be between 3 and 80 characters, got 2.
            floatPrice:
              summary: Money sent as a float
              value:
                type: https://kiezmarkt.example/problems/validation-failed
                title: Validation failed
                status: 400
                detail: priceCents must be an integer number of cents, got 25.99.
            radiusWithoutPostcode:
              value:
                type: https://kiezmarkt.example/problems/radius-without-postcode
                title: Radius filter needs a postcode
                status: 400
                detail: radiusKm was supplied without postcode.
            badCursor:
              value:
                type: https://kiezmarkt.example/problems/invalid-cursor
                title: Invalid cursor
                status: 400
                detail: The cursor is malformed or no longer valid.
    NotFound:
      description: No such resource, or the caller may not see it.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            notFound:
              value:
                type: https://kiezmarkt.example/problems/not-found
                title: Not found
                status: 404
                detail: No listing with id lst_000000.
    Gone:
      description: The listing exists but is `deleted`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            gone:
              value:
                type: https://kiezmarkt.example/problems/listing-deleted
                title: Listing deleted
                status: 410
                detail: Listing lst_5e8802 was deleted on 2026-08-20T10:02:11Z.
    Conflict:
      description: The request contradicts the resource's current state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            terminalStatus:
              value:
                type: https://kiezmarkt.example/problems/terminal-status
                title: Listing is deleted
                status: 409
                detail: Listing lst_5e8802 is deleted and cannot be modified.
    UnprocessableEntity:
      description: |
        Syntactically valid, but rejected by a domain rule: an unknown
        `categoryId`, a top-level `categoryId` on a listing (invariant 4), an
        unknown `sellerId`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            parentCategory:
              value:
                type: https://kiezmarkt.example/problems/category-not-a-leaf
                title: Category is not a leaf
                status: 422
                detail: elektronik has children and cannot carry listings directly.
  schemas:
    MoneyCents:
      title: MoneyCents
      description: |
        An amount of money as an integer number of cents. Never a float, at any
        layer (invariant 3 in `domain.md`). `currency` is carried separately.
      type: integer
      format: int64
      minimum: 0
      examples: [42000, 0, 2500]
    Postcode:
      title: Postcode
      description: Five-digit German postcode. Leading zeros are significant, so this is a string.
      type: string
      pattern: '^[0-9]{5}$'
      examples: ['10437', '12047', '13355']
    Currency:
      title: Currency
      description: ISO 4217 code. Always `EUR` today.
      type: string
      pattern: '^[A-Z]{3}$'
      default: EUR
      examples: [EUR]
    ListingId:
      type: string
      minLength: 1
      maxLength: 64
      examples: [lst_9f2c41]
    SellerId:
      type: string
      minLength: 1
      maxLength: 64
      examples: [sel_0142]
...
    CategoryId:
      description: |
        Current form, dot-separated, at most two levels: `elektronik`,
        `elektronik.audio`. The pre-2024 numeric id is `legacyId` and is never
        accepted here.
      type: string
      pattern: '^[a-z0-9]+(\.[a-z0-9]+)?$'
      examples: [elektronik.audio, haushalt.moebel]
    ImageId:
      type: string
      minLength: 1
      maxLength: 64
      examples: [img_3a1f]
    ListingStatus:
      title: ListingStatus
      description: |
        Lifecycle status. `deleted` is terminal. Only `published` is visible in
        search. Legal transitions are documented on
        `POST /listings/{id}/status`.
      type: string
      enum:
        - draft
        - published
        - paused
        - expired
        - deleted
    Listing:
      title: Listing
      description: |
        A listing as served by the API. There is no seller email here and there
        never will be (invariant 6); a seller is `sellerId`.
      type: object
      additionalProperties: false
      required:
        - id
        - sellerId
        - title
        - description
        - categoryId
        - priceCents
        - currency
        - status
        - postcode
        - createdAt
        - publishedAt
        - imageIds
      properties:
        id:
          allOf: [{ $ref: '#/components/schemas/ListingId' }]
          description: Opaque, server-assigned.
          readOnly: true
        sellerId:
          allOf: [{ $ref: '#/components/schemas/SellerId' }]
          description: |
            The owner. The only seller information in this schema. Any
            human-readable name is resolved separately through `SellerPublic`.
        title:
          type: string
          minLength: 3
          maxLength: 80
        description:
          type: string
          maxLength: 4000
        categoryId:
          allOf: [{ $ref: '#/components/schemas/CategoryId' }]
          description: A leaf category. Top-level categories carry no listings.
        # KNOWN LIE. Declared as a non-nullable integer here, and the import feed
        # delivers `priceCents: null` for listings posted without a price
        # ("Zu verschenken") - 7 of the 120 seeded listings. The schema is wrong and
        # is left wrong on purpose: fixing the type is out of scope, handling the
        # value is not. Any generated client that trusts this declaration will
        # deserialize those seven rows into a crash or a silent zero. See the
        # `unpricedFromFeed` example on `GET /listings/{id}`, and the corresponding
        # note in `docs/project/domain.md` (Where the domain is deliberately silent).
        priceCents:
          allOf: [{ $ref: '#/components/schemas/MoneyCents' }]
          description: |
            Asking price in integer cents. `0` is a legal value and `domain.md`
            records that it is ambiguous in the current data: some `0` rows mean
            free, some mean unpriced. This contract does not disambiguate them.
        currency:
          $ref: '#/components/schemas/Currency'
        status:
          allOf: [{ $ref: '#/components/schemas/ListingStatus' }]
          description: Changed only through `POST /listings/{id}/status`.
          readOnly: true
        postcode:
          $ref: '#/components/schemas/Postcode'
        createdAt:
          type: string
          format: date-time
          description: UTC.
          readOnly: true
        publishedAt:
          type: [string, 'null']
          format: date-time
          description: |
            UTC. Null while the listing has never been published. Set exactly
            once, on the first transition into `published` (invariant 2), and
            never rewritten by a later republish.
          readOnly: true
        imageIds:
          type: array
          description: Ordered. The first entry is the cover image.
          items:
            $ref: '#/components/schemas/ImageId'
          minItems: 0
          maxItems: 20
    ListingCreate:
      title: ListingCreate
      description: |
        Body of `POST /listings`. `status` and `publishedAt` are absent by
        design: a new listing is always a `draft` with `publishedAt: null`.
      type: object
      additionalProperties: false
      required:
        - sellerId
        - title
        - categoryId
        - priceCents
        - postcode
      properties:
        sellerId:
          $ref: '#/components/schemas/SellerId'
        title:
          type: string
          minLength: 3
          maxLength: 80
        description:
          type: string
          maxLength: 4000
          default: ''
        categoryId:
          $ref: '#/components/schemas/CategoryId'
        # Same known lie as `Listing.priceCents`: non-nullable here, null in the feed.
        # Creates through this endpoint must send an integer; the feed does not use it.
        priceCents:
          $ref: '#/components/schemas/MoneyCents'
        currency:
          $ref: '#/components/schemas/Currency'
        postcode:
          $ref: '#/components/schemas/Postcode'
        imageIds:
          type: array
          items:
            $ref: '#/components/schemas/ImageId'
          minItems: 0
          maxItems: 20
          default: []
    ListingPatch:
      title: ListingPatch
      description: |
        Body of `PATCH /listings/{id}`. Every property is optional; absent means
        unchanged. No property is nullable, so `null` is a `400`. `status` is
        deliberately not here.
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 80
        description:
          type: string
          maxLength: 4000
        categoryId:
          $ref: '#/components/schemas/CategoryId'
        priceCents:
          $ref: '#/components/schemas/MoneyCents'
        currency:
          $ref: '#/components/schemas/Currency'
        postcode:
          $ref: '#/components/schemas/Postcode'
        imageIds:
          type: array
          description: Replaces the whole list, including the cover at index 0.
          items:
            $ref: '#/components/schemas/ImageId'
          minItems: 0
          maxItems: 20
    StatusChangeRequest:
      title: StatusChangeRequest
      description: |
        Body of `POST /listings/{id}/status`. `draft` is not a legal target:
        nothing transitions back into `draft`.
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          description: The target status.
          type: string
          enum:
            - published
            - paused
            - expired
            - deleted
        reason:
          description: |
            Free-text marker for who or what triggered the transition. Carried
            through to the status-history row and otherwise unused.
          type: string
          maxLength: 120
    ListingPage:
      title: ListingPage
      description: One page of search results.
      type: object
      additionalProperties: false
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          description: Published listings only.
          items:
            $ref: '#/components/schemas/Listing'
        nextCursor:
          type: [string, 'null']
          description: |
            Cursor for the next page, or `null` on the last page. Opaque: do not
            parse it, do not store it past the result set it belongs to.
...
    Problem:
      title: Problem
      description: |
        RFC 9457 problem details. Every error response in this contract uses this
        shape with media type `application/problem+json`.
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: |
            Stable identifier of the problem kind. Clients branch on this, never
            on `detail`.
          default: about:blank
          examples:
            - https://kiezmarkt.example/problems/illegal-status-transition
        title:
          type: string
          description: Short, human-readable, constant per `type`.
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: Repeats the HTTP status code.
        detail:
          type: string
          description: |
            Human-readable explanation of this occurrence. Free text. Never
            carries a seller email or any other PII.
        instance:
          type: string
          format: uri-reference
          description: URI of the specific occurrence.